# saas-platform/platform-backend/openapi.json
{
  "components": {
    "schemas": {
      "AccountSetupResponse": {
        "description": "Account setup response model.",
        "properties": {
          "account_id": {
            "title": "Account Id",
            "type": "string"
          },
          "message": {
            "title": "Message",
            "type": "string"
          },
          "subscription": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/SubscriptionOut"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "required": [
          "message",
          "account_id"
        ],
        "title": "AccountSetupResponse",
        "type": "object"
      },
      "AccountWithRelationsOut": {
        "description": "Account with subscriptions and instances.",
        "properties": {
          "company_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Company Name"
          },
          "consent_analytics": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Consent Analytics"
          },
          "consent_marketing": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Consent Marketing"
          },
          "created_at": {
            "title": "Created At",
            "type": "string"
          },
          "deleted_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Deleted At"
          },
          "email": {
            "title": "Email",
            "type": "string"
          },
          "full_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Full Name"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "instances": {
            "anyOf": [
              {
                "items": {
                  "additionalProperties": true,
                  "type": "object"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Instances"
          },
          "is_admin": {
            "title": "Is Admin",
            "type": "boolean"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "stripe_customer_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Stripe Customer Id"
          },
          "subscriptions": {
            "anyOf": [
              {
                "items": {
                  "additionalProperties": true,
                  "type": "object"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Subscriptions"
          },
          "updated_at": {
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "id",
          "email",
          "is_admin",
          "status",
          "created_at",
          "updated_at"
        ],
        "title": "AccountWithRelationsOut",
        "type": "object"
      },
      "ActionResult": {
        "description": "Result model for action operations.",
        "properties": {
          "message": {
            "title": "Message",
            "type": "string"
          },
          "success": {
            "title": "Success",
            "type": "boolean"
          }
        },
        "required": [
          "success",
          "message"
        ],
        "title": "ActionResult",
        "type": "object"
      },
...
      "AdminStatusOut": {
        "description": "Admin status response model.",
        "properties": {
          "is_admin": {
            "title": "Is Admin",
            "type": "boolean"
          }
        },
        "required": [
          "is_admin"
        ],
        "title": "AdminStatusOut",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ProvisionResponse": {
        "description": "Response model for provisioning operations.",
        "properties": {
          "api_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Api Url"
          },
          "customer_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "string"
              }
            ],
            "title": "Customer Id"
          },
          "frontend_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Frontend Url"
          },
          "matrix_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Matrix Url"
          },
          "message": {
            "title": "Message",
            "type": "string"
          },
          "success": {
            "title": "Success",
            "type": "boolean"
          }
        },
        "required": [
          "success",
          "message",
          "customer_id"
        ],
        "title": "ProvisionResponse",
        "type": "object"
      },
...
      "SubscriptionOut": {
        "description": "Subscription information output model.",
        "properties": {
          "account_id": {
            "title": "Account Id",
            "type": "string"
          },
          "can_run_instances": {
            "title": "Can Run Instances",
            "type": "boolean"
          },
          "cancelled_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cancelled At"
          },
          "created_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created At"
          },
          "current_period_end": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Current Period End"
          },
          "current_period_start": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Current Period Start"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "max_agents": {
            "title": "Max Agents",
            "type": "integer"
          },
          "max_messages_per_day": {
            "title": "Max Messages Per Day",
            "type": "integer"
          },
          "max_storage_gb": {
            "title": "Max Storage Gb",
            "type": "integer"
          },
          "status": {
            "enum": [
              "active",
              "cancelled",
              "past_due",
              "trialing",
              "paused",
              "incomplete"
            ],
            "title": "Status",
            "type": "string"
          },
          "stripe_customer_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Stripe Customer Id"
          },
          "stripe_subscription_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Stripe Subscription Id"
          },
          "tier": {
            "enum": [
              "free",
              "byok",
              "hobby",
              "pro",
              "enterprise"
            ],
            "title": "Tier",
            "type": "string"
          },
          "trial_days_remaining": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Trial Days Remaining"
          },
          "trial_ends_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Trial Ends At"
          },
          "updated_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Updated At"
          }
        },
        "required": [
          "id",
          "account_id",
          "tier",
          "status",
          "max_agents",
          "max_messages_per_day",
          "max_storage_gb",
          "can_run_instances"
        ],
        "title": "SubscriptionOut",
        "type": "object"
      },
...
      "SyncResult": {
        "description": "Sync operation result model.",
        "properties": {
          "errors": {
            "title": "Errors",
            "type": "integer"
          },
          "synced": {
            "title": "Synced",
            "type": "integer"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          },
          "updates": {
            "items": {
              "$ref": "#/components/schemas/SyncUpdateOut"
            },
            "title": "Updates",
            "type": "array"
          }
        },
        "required": [
          "total",
          "synced",
          "errors",
          "updates"
        ],
        "title": "SyncResult",
        "type": "object"
      },
      "SyncUpdateOut": {
        "description": "Sync update information model.",
        "properties": {
          "instance_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "string"
              }
            ],
            "title": "Instance Id"
          },
          "new_status": {
            "title": "New Status",
            "type": "string"
          },
          "old_status": {
            "title": "Old Status",
            "type": "string"
          },
          "reason": {
            "title": "Reason",
            "type": "string"
          }
        },
        "required": [
          "instance_id",
          "old_status",
          "new_status",
          "reason"
        ],
        "title": "SyncUpdateOut",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "MindRoom Backend",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/my/account": {
      "get": {
        "description": "Get current user's account with subscription and instances.",
        "operationId": "get_current_account_my_account_get",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "title": "Authorization",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AccountWithRelationsOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Current Account"
      }
    },
    "/my/account/admin-status": {
      "get": {
        "description": "Check if current user is an admin.",
        "operationId": "check_admin_status_my_account_admin_status_get",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "title": "Authorization",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminStatusOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Check Admin Status"
      }
    },
    "/my/account/setup": {
      "post": {
        "description": "Setup free tier account for new user.",
        "operationId": "setup_account_my_account_setup_post",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "title": "Authorization",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AccountSetupResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Setup Account"
      }
    },
...
    "/system/instances/{instance_id}/restart": {
      "post": {
        "description": "Restart an instance (provisioner API compatible).",
        "operationId": "restart_instance_provisioner_system_instances__instance_id__restart_post",
        "parameters": [
          {
            "in": "path",
            "name": "instance_id",
            "required": true,
            "schema": {
              "title": "Instance Id",
              "type": "integer"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ActionResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Restart Instance Provisioner"
      }
    },
    "/system/instances/{instance_id}/start": {
      "post": {
        "description": "Start an instance (provisioner API compatible).",
        "operationId": "start_instance_provisioner_system_instances__instance_id__start_post",
        "parameters": [
          {
            "in": "path",
            "name": "instance_id",
            "required": true,
            "schema": {
              "title": "Instance Id",
              "type": "integer"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ActionResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Start Instance Provisioner"
      }
    },
    "/system/instances/{instance_id}/stop": {
      "post": {
        "description": "Stop an instance (provisioner API compatible).",
        "operationId": "stop_instance_provisioner_system_instances__instance_id__stop_post",
        "parameters": [
          {
            "in": "path",
            "name": "instance_id",
            "required": true,
            "schema": {
              "title": "Instance Id",
              "type": "integer"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ActionResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Stop Instance Provisioner"
      }
    },
    "/system/instances/{instance_id}/uninstall": {
      "delete": {
        "description": "Completely uninstall/deprovision an instance.",
        "operationId": "uninstall_instance_system_instances__instance_id__uninstall_delete",
        "parameters": [
          {
            "in": "path",
            "name": "instance_id",
            "required": true,
            "schema": {
              "title": "Instance Id",
              "type": "integer"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ActionResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Uninstall Instance"
      }
    },
    "/system/provision": {
      "post": {
        "description": "Provision a new instance (compatible with customer portal).",
        "operationId": "provision_instance_system_provision_post",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "additionalProperties": true,
                "title": "Data",
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProvisionResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Provision Instance"
      }
    },
    "/system/sync-instances": {
      "post": {
        "description": "Sync instance states between database and Kubernetes cluster.",
        "operationId": "sync_instances_system_sync_instances_post",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SyncResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Sync Instances"
      }
    },
...
  }
}